#8517 ui.part of  plugins/org.jkiss.dbeaver.ext.sqlite moved in plugins/org.jkiss.dbeaver.ext.sqlite.ui
