+++ plugins/org.jkiss.dbeaver.ext.sqlite.ui/OSGI-INF/l10n/bundle.properties
@@ -1,0 +1,2 @@
+Bundle-Vendor = JKISS
+Bundle-Name = DBeaver SQLite UI
