+++ plugins/org.jkiss.dbeaver.ext.sqlite/META-INF/MANIFEST.MF
@@ -4,11 +4,8 @@ Bundle-Name: %Bundle-Name
 Bundle-SymbolicName: org.jkiss.dbeaver.ext.sqlite;singleton:=true
 Bundle-Version: 1.0.80.qualifier
 Bundle-Release-Date: 20200420
-Require-Bundle: org.eclipse.ui,
- org.eclipse.core.runtime,
- org.jkiss.dbeaver.core,
- org.jkiss.dbeaver.ext.generic,
- org.jkiss.dbeaver.ext.generic.ui
+Require-Bundle: org.eclipse.core.runtime,
+ org.jkiss.dbeaver.ext.generic;bundle-version="2.3.117"
 Export-Package: org.jkiss.dbeaver.ext.sqlite,
  org.jkiss.dbeaver.ext.sqlite.model
 Bundle-ActivationPolicy: lazy
